feat: Add align_to_type functions
