+++ src/utils.rs
@@ -2,10 +2,32 @@
 ///
 /// See:
 /// https://doc.rust-lang.org/core/primitive.pointer.html#method.is_aligned_to
-pub fn is_aligned_to(slice: &[u8], align: usize) -> bool {
+pub(crate) fn is_aligned_to(bytes: &[u8], align: usize) -> bool {
     if !align.is_power_of_two() {
         panic!("is_aligned_to: align is not a power-of-two");
     }
 
-    slice.as_ptr() as usize & (align - 1) == 0
+    bytes.as_ptr() as usize & (align - 1) == 0
+}
+
+/// Splits the given `bytes` into padding and a slice that is properly aligned
+/// to `align` bytes.
+///
+/// Returns [`None`] when `bytes` is not large enough.
+pub fn align_to(bytes: &[u8], align: usize) -> Option<(&[u8], &[u8])> {
+    let offset = bytes.as_ptr().align_offset(align);
+
+    if bytes.len() < offset {
+        return None;
+    }
+
+    Some(bytes.split_at(offset))
+}
+
+/// Splits the given `bytes` into padding and a slice that is properly aligned
+/// for `T`.
+///
+/// Returns [`None`] when `bytes` is not large enough.
+pub fn align_to_type<T>(bytes: &[u8]) -> Option<(&[u8], &[u8])> {
+    align_to(bytes, core::mem::align_of::<T>())
 }
